requirements/jenkins.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements/jenkins.txt requirements/base.txt requirements/testing.in requirements/jenkins.in
#
astroid==2.1.0            # via pylint
beautifulsoup4==4.6.3     # via webtest
certifi==2018.11.29
chardet==3.0.4
click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
cssselect==1.0.3          # via pyquery
django-appconf==1.0.2
django-axes==4.5.0
django-choices==1.6.1
django-cors-middleware==1.3.1
django-filter==2.0.0
django-ipware==2.1.0
django-jenkins==0.110.0
django-webtest==1.9.4
django==2.1.4
djangorestframework-camel-case==0.2.0
djangorestframework==3.9.0
drf-nested-routers==0.91
drf-yasg==1.12.0
factory-boy==2.11.1
faker==1.0.1              # via factory-boy
freezegun==0.3.11
gemma-zds-client==0.8.1
idna==2.8
inflection==0.3.1
iso-639==0.4.5
isodate==0.6.0
isort==4.3.4
itypes==1.1.0
jinja2==2.10
lazy-object-proxy==1.3.1  # via astroid
lxml==4.2.5               # via pyquery
markupsafe==1.1.0
mccabe==0.6.1             # via pylint
mock==2.0.0
oyaml==0.7
pbr==5.1.1                # via mock
pep8==1.7.1
pip-tools==3.2.0
psycopg2==2.7.6.1
pyjwt==1.7.1
pylint==2.2.2
pyquery==1.4.0
python-dateutil==2.7.5    # via faker, freezegun
python-dotenv==0.10.1
pytz==2018.7
pyyaml==3.13
raven==6.10.0
requests==2.21.0
ruamel.yaml==0.15.81
six==1.12.0
tblib==1.3.2
text-unidecode==1.2       # via faker
unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
waitress==1.1.0           # via webtest
webob==1.8.4              # via webtest
webtest==2.0.32
wrapt==1.10.11            # via astroid
zds-schema==0.17.10
